Move reference_widget implementations to own cpp
