+++ lib/reig.cpp
@@ -4,93 +4,15 @@
 #define STBTT_STATIC
 
 #include "reig/stb_truetype.h"
+#include "reig/internal.h"
 #include <sstream>
 #include <memory>
 
 using std::unique_ptr;
 using std::vector;
 
-namespace internal {
-    using namespace reig;
-    using namespace primitive;
-
-    template<typename T>
-    bool is_between(T val, T min, T max) {
-        return val > min && val < max;
-    }
-
-    bool is_boxed_in(Point const& pt, Rectangle const& box) {
-        return is_between(pt.x, box.x, box.x + box.width) && is_between(pt.y, box.y, box.y + box.height);
-    }
-
-    template<typename T>
-    T clamp(T val, T min, T max) {
-        return val < min ? min :
-               val > max ? max :
-               val;
-    }
-
-    template<typename T>
-    T min(T a, T b) {
-        return a < b ? a : b;
-    }
-
-    template<typename T>
-    T max(T a, T b) {
-        return a > b ? a : b;
-    }
-
-    template<typename T>
-    T abs(T a) {
-        return a < 0 ? -a : a;
-    }
-
-    template<typename T>
-    int sign(T a) {
-        return a < 0 ? -1 :
-               a > 0 ? 1 :
-               0;
-    }
-
-    primitive::Color get_yiq_contrast(primitive::Color color) {
-        using namespace primitive::colors::literals;
-
-        uint32_t y = (299u * color.red.val + 587 * color.green.val + 114 * color.blue.val) / 1000;
-        return y >= 128
-               ? Color{0_r, 0_g, 0_b, 255_a}
-               : Color{255_r, 255_g, 255_b};
-    }
-
-    Color lighten_color_by(Color color, uint8_t delta) {
-        uint8_t max = 255u;
-        color.red.val < max - delta ? color.red.val += delta : color.red.val = max;
-        color.green.val < max - delta ? color.green.val += delta : color.green.val = max;
-        color.blue.val < max - delta ? color.blue.val += delta : color.blue.val = max;
-        return color;
-    }
-
-    Rectangle decrease_box(Rectangle aRect, int by) {
-        int moveBy = by / 2;
-        aRect.x += moveBy;
-        aRect.y += moveBy;
-        aRect.width -= by;
-        aRect.height -= by;
-        return aRect;
-    }
-
-    template <typename R, typename T, typename = std::enable_if_t<std::is_integral_v<R> && std::is_integral_v<R>>>
-    R integral_cast(T t) {
-        auto r = static_cast<R>(t);
-        if (r != t || (std::is_signed_v<T> != std::is_signed_v<R> && ((t < T{}) != r < R{}))) {
-            std::stringstream ss;
-            ss << "Bad integral cast from " << typeid(T).name() << "(" << t << ") to type " << typeid(R).name();
-            throw std::range_error(ss.str());
-        }
-        return r;
-    };
-}
-
 using namespace reig::primitive;
+namespace internal = reig::internal;
 
 void reig::Context::set_render_handler(RenderHandler renderHandler) {
     mRenderHandler = renderHandler;
@@ -274,291 +196,6 @@ void reig::detail::Window::expand(Rectangle& aBox) {
 
 void reig::Context::fit_rect_in_window(Rectangle& rect) {
     mCurrentWindow.expand(rect);
-}
-
-bool reig::reference_widget::button::draw(reig::Context& ctx) const {
-    Rectangle box = this->mBoundingBox;
-    ctx.fit_rect_in_window(box);
-
-    // Render button outline first
-    Color outlineCol = internal::get_yiq_contrast(mBaseColor);
-    ctx.render_rectangle(box, outlineCol);
-
-    Color color = this->mBaseColor;
-    // if cursor is over the button, highlight it
-    if (internal::is_boxed_in(ctx.mouse.get_cursor_pos(), box)) {
-        color = internal::lighten_color_by(color, 50);
-    }
-
-    // see, if clicked the inner part of button
-    box = internal::decrease_box(box, 4);
-    bool clickedInBox = internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), box);
-
-    // highlight even more, if clicked
-    if (ctx.mouse.leftButton.is_pressed() && clickedInBox) {
-        color = internal::lighten_color_by(color, 50);
-    }
-
-    // render the inner part of button
-    ctx.render_rectangle(box, color);
-    // render button's title
-    ctx.render_text(mTitle, box);
-
-    return ctx.mouse.leftButton.is_clicked() && clickedInBox;
-}
-
-
-bool reig::reference_widget::textured_button::draw(reig::Context& ctx) const {
-    Rectangle box = this->mBoundingBox;
-    ctx.fit_rect_in_window(box);
-
-    bool clickedInBox = internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), box);
-    bool hoveredOnBox = internal::is_boxed_in(ctx.mouse.get_cursor_pos(), box);
-
-    if((ctx.mouse.leftButton.is_pressed() && clickedInBox) || hoveredOnBox) {
-        ctx.render_rectangle(box, mBaseTexture);
-    } else {
-        ctx.render_rectangle(box, mHoverTexture);
-    }
-
-    box = internal::decrease_box(box, 8);
-    ctx.render_text(mTitle, box);
-
-    return ctx.mouse.leftButton.is_clicked() && clickedInBox;
-}
-
-void reig::reference_widget::label::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = this->mBoundingBox;
-    ctx.fit_rect_in_window(boundingBox);
-    ctx.render_text(mTitle, boundingBox);
-}
-
-void render_widget_frame(reig::Context& ctx, Rectangle& boundingBox, Color const& baseColor) {
-    Color frameColor = internal::get_yiq_contrast(baseColor);
-    ctx.render_rectangle(boundingBox, frameColor);
-    boundingBox = internal::decrease_box(boundingBox, 4);
-    ctx.render_rectangle(boundingBox, baseColor);
-}
-
-struct SliderValues {
-    float min = 0.0f;
-    float max = 0.0f;
-    float value = 0.0f;
-    int offset = 0;
-    int valuesNum = 0;
-};
-
-SliderValues prepare_slider_values(float aMin, float aMax, float aValue, float aStep) {
-    float min = internal::min(aMin, aMax);
-    float max = internal::max(aMin, aMax);
-    return SliderValues{
-            min, max, internal::clamp(aValue, min, max),
-            static_cast<int>((aValue - min) / aStep),
-            static_cast<int>((max - min) / aStep + 1)
-    };
-};
-
-void size_slider_cursor(float& coord, float& size, int valuesNum, int offset) {
-    size /= valuesNum;
-    if(size < 1) size = 1;
-    coord += offset * size;
-}
-
-void progress_slider_value(float mouseCursorCoord, float cursorSize, float cursorCoord,
-                           float min, float max, float step, float& value) {
-    auto halfCursorW = cursorSize / 2;
-    auto distanceToMouseCoord = mouseCursorCoord - cursorCoord - halfCursorW;
-
-    if (internal::abs(distanceToMouseCoord) > halfCursorW) {
-        value += static_cast<int>(distanceToMouseCoord / cursorSize) * step;
-        value = internal::clamp(value, min, max);
-    }
-}
-
-enum class SliderOrientation {
-    HORIZONTAL, VERTICAL
-};
-
-SliderOrientation calculate_slider_orientation(const Rectangle& rect) {
-    return rect.height > rect.width
-           ? SliderOrientation::VERTICAL
-           : SliderOrientation::HORIZONTAL;
-}
-
-bool reig::reference_widget::slider::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = {mBoundingBox};
-    ctx.fit_rect_in_window(boundingBox);
-
-    render_widget_frame(ctx, boundingBox, mBaseColor);
-
-    auto [min, max, value, offset, valuesNum] = prepare_slider_values(mMin, mMax, mValueRef, mStep);
-
-    SliderOrientation orientation = calculate_slider_orientation(boundingBox);
-
-    // Render the cursor
-    auto cursorBox = internal::decrease_box(boundingBox, 4);
-    if(orientation == SliderOrientation::HORIZONTAL) {
-        size_slider_cursor(cursorBox.x, cursorBox.width, valuesNum, offset);
-    } else {
-        size_slider_cursor(cursorBox.y, cursorBox.height, valuesNum, offset);
-    }
-
-    auto cursorColor = internal::get_yiq_contrast(mBaseColor);
-    if (internal::is_boxed_in(ctx.mouse.get_cursor_pos(), cursorBox)) {
-        cursorColor = internal::lighten_color_by(cursorColor, 50);
-    }
-    ctx.render_rectangle(cursorBox, cursorColor);
-
-    if (ctx.mouse.leftButton.is_pressed() && internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), boundingBox)) {
-        if(orientation == SliderOrientation::HORIZONTAL) {
-            progress_slider_value(ctx.mouse.get_cursor_pos().x, cursorBox.width, cursorBox.x, min, max, mStep, value);
-        } else {
-            progress_slider_value(ctx.mouse.get_cursor_pos().y, cursorBox.height, cursorBox.y, min, max, mStep, value);
-        }
-    } else if (ctx.mouse.get_scrolled() != 0 && internal::is_boxed_in(ctx.mouse.get_cursor_pos(), boundingBox)) {
-        value += static_cast<int>(ctx.mouse.get_scrolled()) * mStep;
-        value = internal::clamp(value, min, max);
-    }
-
-    if (mValueRef != value) {
-        mValueRef = value;
-        return true;
-    } else {
-        return false;
-    }
-}
-
-void size_scrollbar_cursor(float& coord, float& size, float step, int offset, float viewSize) {
-    float scale = size / viewSize;
-    size *= scale;
-    if(size < 1) size = 1;
-    coord += offset * step * scale;
-}
-
-bool reig::reference_widget::scrollbar::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = {mBoundingBox};
-    ctx.fit_rect_in_window(boundingBox);
-
-    render_widget_frame(ctx, boundingBox, mBaseColor);
-
-    auto step = ctx.get_font_height() / 2.0f;
-    auto [min, max, value, offset, valuesNum] = prepare_slider_values(0.0f, mViewSize - boundingBox.height, mValueRef, step);
-
-    SliderOrientation orientation = calculate_slider_orientation(boundingBox);
-
-    // Render the cursor
-    auto cursorBox = internal::decrease_box(boundingBox, 4);
-    if(orientation == SliderOrientation::HORIZONTAL) {
-        size_scrollbar_cursor(cursorBox.x, cursorBox.width, step, offset, mViewSize);
-    } else {
-        size_scrollbar_cursor(cursorBox.y, cursorBox.height, step, offset, mViewSize);
-    }
-
-    auto cursorColor = internal::get_yiq_contrast(mBaseColor);
-    if (internal::is_boxed_in(ctx.mouse.get_cursor_pos(), cursorBox)) {
-        cursorColor = internal::lighten_color_by(cursorColor, 50);
-    }
-    ctx.render_rectangle(cursorBox, cursorColor);
-
-    if (ctx.mouse.leftButton.is_pressed() && internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), boundingBox)) {
-        if(orientation == SliderOrientation::HORIZONTAL) {
-            progress_slider_value(ctx.mouse.get_cursor_pos().x, cursorBox.width, cursorBox.x, min, max, step, value);
-        } else {
-            progress_slider_value(ctx.mouse.get_cursor_pos().y, cursorBox.height, cursorBox.y, min, max, step, value);
-        }
-    } else if (ctx.mouse.get_scrolled() != 0 && internal::is_boxed_in(ctx.mouse.get_cursor_pos(), boundingBox)) {
-        value += static_cast<int>(ctx.mouse.get_scrolled()) * step;
-        value = internal::clamp(value, min, max);
-    }
-
-    if (mValueRef != value) {
-        mValueRef = value;
-        return true;
-    } else {
-        return false;
-    }
-}
-
-bool reig::reference_widget::textured_slider::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = this->mBoundingBox;
-    ctx.fit_rect_in_window(boundingBox);
-
-    // Render slider's base
-    ctx.render_rectangle(boundingBox, mBaseTexture);
-
-    auto [min, max, value, offset, valuesNum] = prepare_slider_values(mMin, mMax, mValueRef, mStep);
-
-    // Render the cursor
-    auto cursorBox = internal::decrease_box(boundingBox, 8);
-    auto orientation = calculate_slider_orientation(boundingBox);
-    if(orientation == SliderOrientation::HORIZONTAL) {
-        size_slider_cursor(cursorBox.x, cursorBox.width, valuesNum, offset);
-    } else {
-        size_slider_cursor(cursorBox.y, cursorBox.height, valuesNum, offset);
-    }
-    ctx.render_rectangle(cursorBox, mCursorTexture);
-
-    if (ctx.mouse.leftButton.is_pressed() && internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), boundingBox)) {
-        if(orientation == SliderOrientation::HORIZONTAL) {
-            progress_slider_value(ctx.mouse.get_cursor_pos().x, cursorBox.width, cursorBox.x, min, max, mStep, value);
-        } else {
-            progress_slider_value(ctx.mouse.get_cursor_pos().y, cursorBox.height, cursorBox.y, min, max, mStep, value);
-        }
-    } else if (ctx.mouse.get_scrolled() != 0 && internal::is_boxed_in(ctx.mouse.get_cursor_pos(), boundingBox)) {
-        value += static_cast<int>(ctx.mouse.get_scrolled()) * mStep;
-        value = internal::clamp(value, min, max);
-    }
-
-    if (mValueRef != value) {
-        mValueRef = value;
-        return true;
-    } else {
-        return false;
-    }
-}
-
-bool reig::reference_widget::checkbox::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = this->mBoundingBox;
-    ctx.fit_rect_in_window(boundingBox);
-
-    render_widget_frame(ctx, boundingBox, mBaseColor);
-
-    // Render check
-    if (mValueRef) {
-        boundingBox = internal::decrease_box(boundingBox, 4);
-        Color contrastColor = internal::get_yiq_contrast(mBaseColor);
-        ctx.render_rectangle(boundingBox, contrastColor);
-    }
-
-    // True if state changed
-    if (ctx.mouse.leftButton.is_clicked() && internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), boundingBox)) {
-        mValueRef = !mValueRef;
-        return true;
-    } else {
-        return false;
-    }
-}
-
-bool reig::reference_widget::textured_checkbox::draw(reig::Context& ctx) const {
-    Rectangle boundingBox = this->mBoundingBox;
-    ctx.fit_rect_in_window(boundingBox);
-
-    // Render checkbox's base
-    ctx.render_rectangle(boundingBox, mBaseTexture);
-
-    // Render check
-    if (mValueRef) {
-        boundingBox = internal::decrease_box(boundingBox, 8);
-        ctx.render_rectangle(boundingBox, mCheckTexture);
-    }
-
-    // True if state changed
-    if (ctx.mouse.leftButton.is_clicked() && internal::is_boxed_in(ctx.mouse.leftButton.get_clicked_pos(), boundingBox)) {
-        mValueRef = !mValueRef;
-        return true;
-    } else {
-        return false;
-    }
 }
 
 void reig::Context::render_text(char const* ch, Rectangle aBox) {
